+++ include/rocksdb/perf_level.h
@@ -13,14 +13,14 @@ namespace rocksdb {
 
 // How much perf stats to collect. Affects perf_context and iostats_context.
 
-enum PerfLevel : char {
-  kUninitialized = -1,            // unknown setting
-  kDisable = 0,                   // disable perf stats
-  kEnableCount = 1,               // enable only count stats
-  kEnableTimeExceptForMutex = 2,  // Other than count stats, also enable time
+enum PerfLevel : unsigned char {
+  kUninitialized = 0,             // unknown setting
+  kDisable = 1,                   // disable perf stats
+  kEnableCount = 2,               // enable only count stats
+  kEnableTimeExceptForMutex = 3,  // Other than count stats, also enable time
                                   // stats except for mutexes
-  kEnableTime = 3,                // enable count and time stats
-  kOutOfBounds = 4                // N.B. Must always be the last value!
+  kEnableTime = 4,                // enable count and time stats
+  kOutOfBounds = 5                // N.B. Must always be the last value!
 };
 
 // set the perf stats level for current thread